std.heap.ArenaAllocator: make State.promote more safe
Previously, std.heap.ArenaAllocator.State.promote required the user to
manually copy the arena state back to the original pointer when done
with it. The original state also couldn't be used elsewhere until it was
copied back. This was a clear footgun, as it's a very easy detail to
miss and forgetting to do it would likely cause strange memory
corruption issues which could be hard to track down. Now, promote
returns a new structure containing a pointer to the original state, and
this pointer is mutated whenever an operation is performed through its
allocator, removing this footgun.

To have a nice API, the returned Promoted struct can be referenced only
through const pointers, since the struct itself never needs to be
mutated. This allows the direct usage 'state.promote().allocator()' to
work safely. However, previously, the std.mem.Allocator API didn't allow
a way to do this without resorting to hacks like @ptrToInt, as the
type-erased pointer was stored internally as a *anyopaque. Therefore,
the definition of Allocator has also been modified to allow storing
const pointers through a union. This required a few changes to std, but
nothing should be hugely breaking.

Resolves: #13409