langref: fix the slices.zig doctest
In the slices.zig doctest, the code
"const all_together_slice = all_together[0..]" is incorrect, since the
actual type is a pointer to an array, instead of a slice.

Use a runtime-know value to slice the array.

In the next "slice pointer" test, clarify that slicing a slice to
produce a pointer to an array, requires comptime-know indexes, not just
constant indexes.