Fixes #12685.
The problem was that paths in the cache are assumed to endwithout the path separator, but are added verbatim from the command line
in several places. The solution here is to make sure that all prefixes are without path separator by cutting it off when it is passed
to Cache.addPrefix. This way, the assumption is met.