lib: add panic test runner
Panic test runner starts itself as child process with the test function
index representing a test block. The custom panic handler compares a
global previously written string.  If the panic message is expected, the
success count is increased and the next test block is run. On failure
the parent process prints collected results, before it terminates.

This means, that only one @panic is possible within a test block and that
no follow-up code after the @panic in the test block can be run.

Depends on #14152.
Closes #1356.