doctest: make the "Exporting a C Library" example work
The mathtest example of exporting a library with the C ABI it not
working, since the mathtest.h file is not generated.

Update the shell sample, adding the -femit-h flag to the zig build-lib
command.

Show an example of how to copy the zig.h file to the project directory,
using zig env and jq.  I'm not sure if there is a better/official
method.